Added super admin + small bug fix
At each server start -> check if there is a super admin -> if not, creates one and writes the password to the terminal -> if more than 1, then warning.
Added super admin to author card, and to all of the other places it's useful.
Also, I fixed the bug that response.status was equal to 200 instead of 204.