Created /blog route (still doesn't handle new articles) + other stuff
Changed image on display for article to image upload instead of entering url.
Also, small bug fix: when importing an image too big or another kind of file (-> error) the text next to the "browse" button was displaying the name of the file even though it shouldn't.
Created BlogRoute ('/blog')
Also tested having 20, 000 characters in the article page.